refactor : change method signiture ans folder check output
